+++ 3d-viewer/trackball.h
@@ -47,8 +47,7 @@
  * The resulting rotation is returned as a quaternion rotation in the
  * first paramater.
  */
-void
-trackball(float q[4], float p1x, float p1y, float p2x, float p2y);
+void trackball(double q[4], double p1x, double p1y, double p2x, double p2y);
 
 /*
  * Given two quaternions, add them together to get a third quaternion.
@@ -58,21 +57,18 @@ trackball(float q[4], float p1x, float p1y, float p2x, float p2y);
  * rotation, the second and third the total rotation (which will be
  * over-written with the resulting new total rotation).
  */
-void
-add_quats(float *q1, float *q2, float *dest);
+void add_quats(double *q1, double *q2, double *dest);
 
 /*
  * A useful function, builds a rotation matrix in Matrix based on
  * given quaternion.
  */
-void
-build_rotmatrix(float m[4][4], float q[4]);
+void build_rotmatrix(GLfloat m[4][4], double q[4]);
 
 /*
  * This function computes a quaternion based on an axis (defined by
  * the given vector) and an angle about which to rotate.  The angle is
  * expressed in radians.  The result is put into the third argument.
  */
-void
-axis_to_quat(float a[3], float phi, float q[4]);
+void axis_to_quat(double a[3], double phi, double q[4]);
